+++ src/main/java/usay/app/domain/member/entity/Member.java
@@ -12,8 +12,8 @@ import java.util.List;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import usay.app.common.entity.BaseEntity;
 import usay.app.domain.chat.entity.Chat;
+import usay.app.domain.common.BaseEntity;
 
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
